pyproject.toml
[project]
name = "shopify-st"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiodataloader>=0.4.2",
    "ariadne>=0.26.2",
    "ariadne-auth>=0.1.1",
    "ariadne-graphql-modules>=0.8.0",
    "asyncpg>=0.31.0",
    "google-cloud-pubsub>=2.23.0",
    "litestar[pydantic,sqlalchemy,standard]>=2.18.0",
    "sniffio>=1.3.1",
    "uvicorn[standard]>=0.38.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.12.0",
]
